CAP-55 ふらふら中に攻撃されたらコンボ数を加算してふらふら時間を延長する

+++ src/ccm/Collision/AttackCollisionActor.cs
@@ -14,8 +14,8 @@ namespace ccm.Collision
         public int Power { get; set; }
 
         /// <summary>
-        /// よろけ値
+        /// よろけ倍率
         /// </summary>
-        public int Shock { get; set; }
+        public float Shock { get; set; }
     }
 }
